+++ test/container_type_test.py
@@ -30,11 +30,15 @@ class ContainerRobotizeTestCase(unittest.TestCase):
         for ref in ["c5", "C5", "14", 14]:
             self.assertEqual(14, dummy_type.robotize(ref))
             self.assertEqual((2, 4), dummy_type.decompose(ref))
+        self.assertEqual([0, 5], dummy_type.robotize(["A1", "B1"]))
 
     def test_robotize_decompose_checks(self):
-        for ref in [["A1", "B1"], 1.0]:
+        for ref in [1.0, 4.3]:
             with self.assertRaises(TypeError):
                 dummy_type.robotize(ref)
+            with self.assertRaises(TypeError):
+                dummy_type.decompose(ref)
+        for ref in [["A1", "B1"]]:
             with self.assertRaises(TypeError):
                 dummy_type.decompose(ref)
         for ref in ["D1", "A6", 15, "2A2"]:
@@ -48,6 +52,10 @@ class ContainerRobotizeTestCase(unittest.TestCase):
         self.assertEqual("A5", dummy_type.humanize(4))
         self.assertEqual("B2", dummy_type.humanize(6))
         self.assertEqual("C5", dummy_type.humanize(14))
+        self.assertEqual(["A1", "B2"], dummy_type.humanize([0, 6]))
+        self.assertEqual(["A1", "B2"], dummy_type.humanize(["0", "6"]))
+        with self.assertRaises(TypeError):
+            dummy_type.humanize("0.1")
         with self.assertRaises(ValueError):
             dummy_type.humanize(15)
         with self.assertRaises(TypeError):