+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 Added
  - discard() and and set_storage() methods for containers
  - gel_purify instruction to instruction.py and protocol.py
+ - support for list for container_type.robotize, container_type.humanize
 Changed
 
 Removed